backend/mcp/requirements.txt
mcp
psycopg2-binary
fastapi
uvicorn[standard]
pydantic
